Checkpoint: Adding social networks to the share dialog. Facebook is mostly done, mostly tested, but it can't really be tested until they can connect to our backend.
We now route all "referral" urls through welcome, even if you
are a guest. In that case, the affiliate is "0" and the welcome servlet
doesn't set the affiliate.

Because: Whatever url we have for the share link, facebook loads and
parses. Redirecting to our home page gives a crappy share experience,
so we detect the facebook user agent and look some things up from the
database to return the scene name and snapshot. We do not want to
duplicate this logic (or complicate things) so I think it's totally
reasonable to route everything through /welcome/.

Still to do:
- some serverside translations
- return nice image/description for games.