So just use an Image control. We don't need it to dim.
Fucking TabNavigator: it chops the text on the first tab,
but as soon as you so much as hover the mouse over it,
it wiggles everything into place. I tried programmatically
wiggling it, but I couldn't. I could increase the width
of things until this bug doesn't happen, but the increase
would have to be rather drastic. Punting.