Permaguests can now be created via a game server login * Clicking a PLAY button in GWT with no token will create a permaguest account and use the visitor   id generated in GWT. * Logging in (eg registering) in GWT will wipe out the flash permaguest cookie the next time flash   is opened. * Playing an embedded game (html) with no permaguest flash cookie will create a new permaguest   account with a new flash-generated visitor id and set the flash cookie. * Playing an embedded game with a permguest flash cookie will login as the previously created   permaguest * Bug fix for simpler world-based permaguests: if present, a permaguest flash cookie will be used   when logging into the world server. * Going from an embedded game to whirled and loading a room or playing a game there will pick up   the permaguest progress from the flash cookie, propagating it back to GWT for registration. * Going from an embedded game to the registration page on GWT manually has no way of picking up the   cookie, so the permaguest will be lost. * The upsell link from embedded games that will bring along the permaguest username for registration   is not yet implemented. * Various edge cases to do with opening two browsers at once can cause multiple permaguest accounts   to be created. We are not worrying about those.
As a consequence of these new features, starting a game from outside the world is slower because
the world server login is only initiated after the game server login finishes (regardless of
whether permaguests are enabled on the server). After initial optimism, this turned out to be
difficult to keep in parallel. In addition to displaying world data in the lobby screens, we also
need to be able to do things like track client actions and get fields off the MemberObject when the
game mode is entered. Each requirement will need to be individually addressed before returning to
parallel logins.