Stole another old idea from Playsite: For in-progress tables, add a little info "hotzone" button that can be hovered to instantly pop a tooltip that displays the game options.
It's a little funny, but it's probably better than either assploding
the lobby panel with too much information, or hiding it all
in a tooltip on the player count label, as it was.

I also put the game's custom options first, and put each piece
of info on a new line.