Fixed bug where the thane backend was subscribing to player objects multiple times. I'm not sure if it affected users but was spamming the logs and obscuring real errors.
It turns out isInited was a very confusing thing to override, so got rid of that. Instead, the
base backend now guarantees (as it always used to) the propering ordering of occupant addition,
removal and transition between occupant and player, based only on the OCCUPANT_INFO and PLAYERS.
These are broken into two functions each. One for for when the transition occurs and one for
dispatching to user code.

The thane backend then overrides all of the occupant bottlenecks and makes sure that the ordering
is maintained for the user code, but also that the player objects are available when the dispatch
occurs. Due to corner cases, this required a tracking array of which occupants had been added
(e.g. there would otherwise be no difference between a player being added and a occupant changing
to a player, so if the subscription were to stall for some reason we would not know whether to
dispatch the occupant removal).

Finally, in order to ensure the game is kicked off correctly, readyToStart (renamed to avoid
confusion with isInited) is overridden and checks for subscriptions.

A side effect of this rejigging is that converting a player object to an occupant info is no
longer required so msoy doesn't need to override anything to perform the cast to PlayerObject.