Wonderful flex hackery to not let the name of a table blow-out the layout. Everything else in here is all about relative sizes, percent widths. Good times. But to limit the size of a component you need to specify the maximum pixels. Fuckers.
Honestly, the amount of time we've all wasted dealing with this
pile of shit is outrageous. If we weren't in the middle
of serious business, I'd say it would be time to write
our own layout containers for flex that can lay things out
without sucking. We'd save so much time.