Various simplification and correctness changes relating to permaguests and affiliate handling.
1. Affiliates are now only member ids. They have never been anything other than
member ids and there's no reason to maintain a bunch of complex unused code to
support mapping some arbitrary string when we can just create an affiliate
account for any "official" affiliate before we start a campaign. kLOC--.

2. There are only three ways in which an account is created in Whirled now and
two of them are fundamentally new and how 99.9% of all accounts are created.

   a) you come to the Whirled website and never run Flash and create an account
   directly from GWT (this is how we used to do it)

   b) you come to the Whirled website and you start Flash somewhere and a
   permaguest account is created for you

   c) you use a Flash embed somewhere in the wild and a permaguest account is
   created for you

To properly handle affiliation, we need to communicate the affiliate id from
the Flash client to the server when logging on. This change went into a prior
commit. In this commit we now use that affiliate information to properly
affiliate the permaguest account (and the eventual member that it hopefully
becomes).

We now have to communicate the affiliate information from GWT to Flash in case
'b' so that when a permaguest account is created for you, it immediately has
the correct affiliate information. This change is in this commit.

If you create an account without ever accessing Flash, your affiliate
information is obtained directly from the cookie, as before. This has not
changed.

A few other bits got tidied up along the way, but those are the meaningful
changes.