Restructure the way ItemRecord derived classes define their schema version so that we can compute the proper migration version when something in ItemRecord changes. An ItemRecord derivation's version is made up of:
baseVersion * bsaeMultiplier + itemVersion

This allows us to increase baseVersion when something in ItemRecord changes and
itemVersion when something in the derived class changes. However, we need to do
some jiggery pokery to get the right migration version when something in
ItemRecord changes. For example, we might have itemVersion values of:

GameRecord - 18
PhotoRecord - 1
AvatarRecord - 3

and a baseVersion of 21, which means that the migration version for each of
these classes is:

GameRecord - 21018
PhotoRecord - 21001
AvatarRecord - 21003

In order to compute this we have to reflectively read GameRecord.ITEM_VERSION,
PhotoRecord.ITEM_VERSION, etc. and combine them with
ItemRecord.BASE_SCHEMA_VERSION. So now we have
ItemRepository.getMigrationVersion() to handle precisely this.

Note that this is only necessary when something in ItemRecord changes. If
something changed in GameRecord, say, then you can just increment
GameRecord.ITEM_VERSION and compute the full schema version at the time of
migration (21019 say) and register the migration for just GameRecord.class with
just that version. It's only when ItemRecord changes and we need to
polymorphically register migrations for all ItemRecord derivations with the
correct schema version that this tricky business is required.