Fixed the sidebar chat problem. Well, I sorta just hacked it.
The problem was that with the changes to Prefs that I made,
when you log in like normal you start out with blank prefs
and then transition to using your saved prefs. When this transition
is made, all the saved prefs are dispatched. This is necessary
because by that time the control bar is set up, and otherwise
we'd display the wrong volume.

However, this rapid-fire of preference changes somehow caused
the chat overlay to booooch. Or it was just the fact that it was
trying to set up before it was ready. I had some theories,
but couldn't figure out what was really going on.

Part of the problem is the insane complexity of this. The ChatOverlay
should not be putting itself into a sidebar, it should render chat,
and that's it, period. Something else should be putting it into
or out of sidebar mode. Part of this insanity is why GameChatContainer
exists, I think.

So: wait until the embedded state is known before listening to prefs
changes. That's how it used to work and it seems to be fine.

Also: tore out the TabbedChatDisplay interface. It was a giant NOOP
leftover from the days of yore.

When we have some time, refactoring this chat stuff so that
it's optimized to work the way we settled upon would surely
clean up a lot of things.