Giant cleanup of game entry handling. This vastly simplifies a bunch of byzantine shit that lingered from the half dozen total rethinks of how lobbies work and institutionalizes some new stuff. A few notable things:
1. All games that support single player are started immediately. We'll never
show the lobby first except for multiplayer-only (and by corollary party)
games.

2. We no longer route the game oid through the URL. This was partly a technical
wrinkle but also to end the annoyance of back button breakery when
world-game-p-ID sent you immediately into a game which then changed the URL to
reflect the game oid. We can and should come up with a way to encourage users
to invite people into their already running game and that can use the now much
more robust world-game-p-ID-MID mechanism which says "I want to play game ID
with player MID" and it automatically puts you at their table if they're
matching or in their game if they're playing.

3. Single player games are no longer private by default. This makes going into
someone's running game generally work rather than generally not work. If we
want to reintroduce privacy, we should do it after the fact and put it in the
game menu: "Allow kibbitzing" or something.

4. Joining someone else's table or game should work now rather than being
horribly broken like it was before.

5. We don't rely on a world connection at all in most cases except where you
need to immediately go into a lobby, so getting into the game is faster.

6. A fuckload of code went away. Yay. Code is the enemy.

7. We currently use world-game-p-ID and world-game-i-ID which do the exact same
thing, but -i- means you are arriving from some sort of invite (the URL shared
from the lobby or an invite sent via the "invite people to play with you"
mechanism, or via the invite that comes in through a notification). This is
just a tracking convenience that will allow us to see how many people are
getting into games via invites versus going in by clicking a play button
somewhere.