Handle gameLobby here as well. Started wiring up GOT_EXTERNAL_NAME handler but stopped because it's fucking complicated. Goddamned asynchronicity makes our life a living hell. The following things happen in parallel:
1. We start loading the Kongregate API SWF
1a. When that completes, we ask what our username is and tell MsoyClient
2. We start loading our client SWF
2a. When that completes we logon to the game server
2a1. When we're logged on, we create and enter our game
2b. When that completes we logon to the world server

Any of 1a, 2a1 or 2b could complete first.

So we can't just hope that 1a completes first and then stuff the external name
into our GameCredentials so that the game server knows about our desired
display name. Plus it would be a hack to put display name in our credentials
since we normally don't do any display name stuff there.

We could assume that 2b happens last and then when we tell the world server
about our new display name, it could dispatch a PlayerNodeAction to tell any
game server that happens to be hosting said member that their display name has
changed, that server would then update the player's name in any game they were
playing. This has the nice benefit of working in this case as well as if/when
you change your display name via other means.

However, if 2b happens substantially after 2a1 happens, the game will probably
have already grabbed the "Guest XXX" name from the game object and it would be
possible that the name won't show up in some cases. I'll be damned if I'm going
to push this complexity all the way out to GameControl and dispatch an event to
the game telling it that the player's display name has changed. That's TMFingI.

Probably I'll just do the PlayerNodeEvent and just lump it if 2b takes too
long. It would be nice to sneak in a 2a-pre1 where we set the name on the game
server because it's almost always going to be the case that 1a happens before
2a, but that's a bunch more code and then I'll be really pissed when 1a happens
*after* 2a and we're too late to update our name on the game server anyway and
the whole thing still doesn't work perfectly after writing way too much
goddamned code.

Those little shits on Kongregate never seem to like anything we put up anyway.