Invite your friends from Facebook to play games - now live
Tweaked some text. Switched to POST method for the form action, but no luck, still only 1 user id is reported. Count up the number of user ids reported anyway and log to a new panopticon event type, so that the number will be magically > 1 when Facebook decides to fix this. Log the ids[] query param for now so we can see what Facebook is sending us.