We no longer ever pass visitorId from Flash to GWT. If you're ever in Flash then you create a permaguest account which has the Flash-assigned visitorId saved with it and those permaguest credentials are shipped to GWT if you click anything.
If you create an account from GWT having never touched a Flash client, then
you'll use the visitorId assigned there, but it will never be passed in. We
created it when you show up.

If you come to GWT (and get a visitor id assigned) and then go into Flash,
causing Flash to create a permaguest account, it first calls back into GWT to
get the visitor id and then passes that up to the server which uses it when
creating the account.

So all of our asses should be covered.