Nix A/B test. The results were 9.23% converted, 1.51% retained for the combined page and 8.47% converted and 1.40% retained for the split page, so we go back to the combined page.
Note to future A/B test implementors. Don't use a redirect. Just show the
landing page in question at #landing. Redirecting fucks up Google Analytics and
makes it impossible to tell where people go *after* landing, which is the only
way we can get an idea of how to improve the landing page.

Ideally Google Analytics wouldn't fuck this up, but they do, so we cope.