Jesus Horseradish Christ, finally tracked down a problem with links in text "bleeding over" onto the text afterwards. It's not a bug, it's a feature! If you don't want _any_ property of the TextFormat to bleed over from the one before it in the field, you need to make it non-null.
Note that well kids. I knew that the default would come through
for some properties, but now I know that all properties "bleed through"
even from custom TextFormats set on earlier sequences in the field.

Remember this, or someday you'll waste a bunch of time, too.

So this fixes some stuff with both notifications and chat,
paving the way for some other pending changes on those fronts.
(Included here is some of the jockying for the new chat feature).