Implement an experimental algorithm for improving catalog query cache hit rates, motivated by the following observation(s):  - Viewing a shop page forces the database to enumerate a large dataset and then to sort it, in order to finally return a tiny fraction of it, discarding the remaining 99.9%  - The user will then click the 'next' button, putting the database through precisely the same ordeal, except this time it's a different 0.01% that's kept  - The user will then do this again 40 times because -- who knew? -- shopping for cool stuff on Whirled is apparently fun (also because our search functionality ain't all that just yet)  - Returning the top 100 rated items (or the top 100 of any ordering) is no more expensive than returning the top 20, or items 71-93, or whatever
So the algorithm is pretty obvious: intercept offset/limit queries, round the offset down to, say, the nearest 100, always fetch 100 at a time, finally carve out the portion of that result set that was actually requested and return it.

Thus looking at pages 1, 2, 3, 4 and 5 of the shop or some shop search, will all translate to the same database query, which will result in the cache being consulted for the same collection query each time. At the very least, we've implemented a prefetching algorithm, but hopefully there'll be some other synergetic effects as well.

Spammy logging to be removed as soon as I've tested this stuff a bit on the dev server.