Sigh, the most common reason for a memory rejection is that it's not the avatar you're wearing.
I think wyvern avatars must update memory on other wyvern avatars, so that
was too restrictive. Make it the same as pets: the avatar/pet just has to
be in the room, but memory updates can come from anyone.
Ugh. Perhaps it was a mistake to not require control, at least for actors.