Well, it's my fault for not checking a shop item on dev prior to release. But: we gotta be careful because the studio extends World.
Changed SocialDirector to just assume that the MsoyController is a
WorldController, which it always will be, presently. This can be adjusted,
but guess what?  If and when we do have a separate game client, you can't
just have a reference to WorldController. That will cause it to be compiled in.
No, you should create an interface that WorldController can implement,
and that at worst the GameClient will compile in that interface.
But this shit should probably be done with events instead anyway.
Doesn't LocationDirector do that?