Two policy related bits:
1. On local installations we were trying to load
http://IP/media/crossdomain.xml which doesn't exist. So we need to munge the
media URL to strip off the media/ at the end, which we now do. You might think
this fixes all those annoying security warnings, but you'd be wrong. Now we
have:

2009-03-20 16:43:17,542 INFO: com.threerings.msoy.client.MsoyClient Loading
media policy: http://192.168.54.41:8080/crossdomain.xml

which is correct and which exists and grants permission to the client and
then a little later:

*** Security Sandbox Violation ***
SecurityDomain
'http://bering.sea.earth.threerings.net:8080/clients/0/world-client.swf?' tried
to access incompatible context 'http://192.168.54.41:8080/crossdomain.xml'

The former line used to be /media/crossdomain.xml so somehow it's specifically
related to the policy we're loading, and loading the correct one fails to fix
the problem. Awesome!

2. Avoid trying to reload our world server security policy on logons after the
first. That seems to cause out-freaking.