Figured out WTF #2.
So... let's just take a little trip down Flex Can Eat Shit Lane.

When a Container is disabled, it draws a translucent white "blocker" over
itself. Great. That's a great default thing to do.

Problem: no way to really disable this behavior. The method to create
the blocker is private. I guess the variable itself is mx_internal, so
we could override commitProperties() and be sure to manually destroy
the blocker at the end, every time.

The problem is that the code in Container checks the 'enabled' state directly,
rather than the underlying holder variables. So we can't do the reasonable thing,
which is to tell the superclass that we're enabled while reporting to callers
that we're disabled (basically come up with our own concept of what disabled means-
in this case I want to disable the buttons, not draw a white square over
"get bars".). We can't "disable" the white blocker without reporting to callers
that we're always enabled.