Added the owner name and access control to the RoomObject as well. I already want the owner in the AVRG API and chances are most of the other metadata will one day find cause to be exposed.
Also, this is a small step toward the glorious future in which we ditch the
whole scene model/scene update system for rooms and just publish everything in
the RoomObject. There will be no versioning or fiddly business and we'll be
more efficient with our database accesses. I will much rejoice.