New way of distributing party information. Now each node has 3 DSets for parties: one with PartySummary (immutable), one with PartyInfo (contains all the "public" mutable stuff), and one with MemberParty (maps memberId to partyId).
And so: we have nearly all of the information, everywhere.
We no longer need to go to the invoker to satisfy
PartyBoard requests or PartyDetail requests.
Also removed the NodeAction that updated someone's PartySummary,
as that's now all published and we just watch the node objects.

I can't believe this took all day. Lots of distractions?
It's not quite done yet, either, just gotta finish one small thing.

And: still some cleanup after this
(PartySummary will be removed from MemberLocal). And
including leader information out to PartyPlaceObjects.