Revamped the way we track players in games. The old system failed to count embedded game players as we would try to update their MemberLocation before they had established their world connection and they would then not show up as playing in any of our online counts.
Now we track game locations on the node that's hosting the game which
simplifies a few things and complicates a few others.

Along the way, I cleaned up a bunch of friend handling stuff which was a bit
wasteful with CPU (scanning every node object for a friend's online status for
every of a player's friends instead of obtaining the online status for all
friends at once) and network (updating the online status of all online friends
every time a player switched nodes whether or not it had changed).