More unfucking: when a player does anything in a game session (joins a table or starts an AVR or parlor game), we mark them as "playing" that game by tracking their currently playing game in their PlayerObject (and doing all the broadcasting and blah blah that we did before).
Then we do not clear that information out until they end their session on that
game server. Since the client establishes a specific connection to a game
server to play a specific game, this is exactly what we want. Before we were
manually tracking who was playing games and in tables and trying to set and
clear things as players moved around, but that resulted in all sorts of
brokenness and pointless network traffic when the player went from game to
table to game or any other weird ordering.