Call a spade a spade. On the admin page it says: inviter and invitees. Not so: that was your affiliate and your... somethings. I'll fix that 2nd part in a sec.
Your inviter is always your affiliate, but your affiliate is not always an inviter.
I'm not sure wtf is going on with MemberRepository.loadMembersInvitedBy()
it seems to look up based on affiliate id, but then filter to be only invitees?
We should do one or the other, and let's just do affiliates.