Finished the new deal, which is thus:
- when a client requests /go/... (or /welcome/...) or / we assign them a
visitor id if we think they should have one (they have no cookies that lead us
to believe that they're a returning user)

- /go also parses the vec_XX information, records it directly and omits it from
the redirect URL; this avoids two additional requests from new landers to
report the visitor id they assigned to themselves and the entry vector that is
associated with it, and it makes the URL less ugly; this means that vec_XX will
*only* work in /go or /welcome URLs, but we should always be linking to /go or
/welcome URLs anyway and never to /# URLs

- the old visitor id tracking remains for people who show up with a WHO cookie
for example but really are new, which we only discover after they start
browsing around Whirled and do something like try to register or play a game as
a guest; in that case, we create the visitor id on the client as before and
tell the server that we did so

This was all just annoying me and shouldn't result in an appreciable change to
our numbers. It will mean that people who fail to run JavaScript for any reason
(or close their browser before the JavaScript has downloaded) will be assigned
visitor ids. These people were previously totally invisible to us, so if it's a
meaningful number, it will be good to know.