Game resolution involves coordinating between a bunch of peers and loading things from the database, therefore it falls under that class of things that has heretofore been unnamed, but I christen it "distributed programming-complicated" or DP-complicated for short.
Before we played things fast and loose, now we'll do things (mostly) properly,
which involves keeping lists of penders, handling no fewer than half a dozen
different potential failure cases and taking extra special care to track and
release various resources and table entries.

I also now refuse to resolve a lobby on a server that does not think it is
hosting the game in question. In theory this should never happen, but we live
in the messy real world where shit happens all the damned time. So we cover our
ass and stock up on toilet paper.