The featured place view and the chromeless client are chromeless, so let's just test for chromelessness where we care about whether or not we have chrome and not test for featured placeviewness.
We also now don't add the HeaderBar if we're chromeless but we keep it around
because other things assume it will be there and call into it. The HeaderBar
doesn't listen for location changes and whatnot until it's added to the UI to
ensure that it doesn't crap itself when location changes come in and it
discovers that it has no children.