Revamped the Galaxy panel while I was here. The main Galaxy page should load fast (ie. in one query from the server) and should delegate to another page if you actually want to browse or search groups. This is a better experience for the user (they see the main groups page quickly and its focused and leads them to more detail if they want it) and faster loading and less strain on the server.
Then I added a special groups list panel which is used if you "See all..." or
click any of the categories or do a search. This is a page that is focused on
searching and browsing groups. It uses the entire page, rather than being
crammed in a little box like our old browsing interface was, and it shows the
groups' blurbs because this is *exactly* where the blurb was meant to be
shown. Where a user is searching/browsing groups and wants a brief explanation
of what the group is all about. OMG it's modestly sensible user interface
design. Watch out, it's probably the second sign of the apocalypse.

I died a little inside making galaxy_featured_title_bg.png but I would die even
more fucking around to make a scalable round header. Why does all of our shit
have to be fucking round?