Two main changes:
1. When we noted a new visitor we almost always had the entry vector handy. Now
we always have the entry vector handy, so we now note the new visitor and the
entry vector at the same time.

2. Always come up with some sort of entry vector. If an explicit one is not
provided, create one based on the page they landed on (or in the case of a
vectorless embed landing, which should never happen, use world_session or
game_session).

This tidies up some wackiness with re-reporting entry vectors and it allows us
to compare organic landers in the same way we compare vectored landers.
Everyone came in through somewhere, so let's track and report them all.