More cleanup: we no longer need to stuff the referrer into a cookie and then have GWT regurgitate it to us. We just look it up on the server at the time the client makes their first request and store it along with all the other shit.
If GWT does end up reporting that it created a visitor info (which should
happen rarely if at all and we'll be able to see it happen because those
vistors will be marked with gpage.* entry vectors), we lose the referrer info.
Given that this will essentially never happen, we'll go with this simpler
approach.