Whoever exposed a varargs updateGroup() on GroupRepository should be smacked, the Repository classes are meant to encapsulate persistent modifications, not just throw the doors open to any application level class that wants to fling random crap at the database.
Further, when one comes up against a wrinkle like this, don't nix the fricking
typed method and keep the varargs method and change the one caller of the typed
method to use the varargs method. Change the one caller of the varargs method
to use the typed method (or in this case even better, change it to use a
sensibly named method with a specific purpose and nix the varargs method
completely).