3000 was the old default. This value is supposed to be updated with the persistent set value from the database.
Somehow: the server started up, saw 3000, and ran with it.

But I could fire-up the Dashboard and view the value, and it was 10,000.
But the exchange saw 3000. I don't fucking know, I blame mdb's database
tweaking.

So this just changes the default in code. This doesn't fix the problem,
this hides the problem. "Nothing to see here", we'll tell our users,
"Everything is fine".