After thinking about it some more, change to this:
- Have the billing server tell us the number of months the subscriber
will be a subscriber before it gets back to us.
- record a boolean 'subscriber' in the record, instead of relying on the date.

Do that instead of having it tell us an explicit end date so that we
avoid any funny business with
- time trading
- us granting every subscriber some bonus days, due to system downtime or
something

So now we simply decrement their "barGrantsLeft" every month and give them
some bars. Also added fields for the lastSpecialItem.
When they're made a subscriber, we'll note the current specialItem and give
it to them if they hadn't already received it. Whenever a new special item
is made, we'll give it to all current subscribers.