It's not OK for ItemRepository to have a dependency on MoneyExchange. Repositories should have no dependencies, or if absolutely necessary they can depend on other repositories. They should definitely not depend on something that sucks in all of Presents and basically renders the repository completely useless outside of a fully operational Death Star.
Of course, nothing is easy which is why this hack was no doubt just slipped in
in the first place. CatalogLogic now handles loading listings from the catalog
and it handles the amazing four hundred thousand variations and random query
bits that are involved in loading and sorting the catalog. Yay for complexity.