The flex RSL files (Runtime shared library).
The swz file is the signed one that is cached inside the flash cache.
The flash plugin will not even ask the browser if it has a matching
swz already downloaded.

The swf is the same stuff, but as an unsigned fallback swf for older
flash players, that AFAIK we don't support right now anyway.