Arrrgh.
We have a number of small helper swfs that don't use flex, but DO do things
like embed an image, which ends up being a mx.core.BitmapAsset, which
wasn't being compiled in due to us excluding the library code.
So: include the necessary classes. These swfs are still very small, as only
a few tiny classes/interfaces from flex are dragged in.