So, we need to specify the full URL for our RSLs otherwise when we load into an embedstub the stub's url base is being used to qualify these relative urls. On SOME browsers.
(WTF? Isn't the flash plugin telling the browser what to download?
Why would the flash plugin do one thing on IE8 and another on IE7?
Giant fuck yous to: Adobe, Microsoft, and anyone else standing nearby.)

So we now move the config file to an *.in file and transform it prior
to building.