Ugh, I had an inkling that this server FBML malarky was a bit too easy... well, now the code better reflects reality and will actually work with upcoming #facebook-invite page.
For the masochistic: Most FBML display tags can be used in an iframe without the server fbml wrapper, so the test code did not even need it. The serverfbml tag is used for things that require user input. Apparently, it *must* be wrapped with <script> and <fb:fbml>. If not, it just fails silently with no log messages or exceptions. Hurray for undebuggable code! But wait! There's more! GWT DOM (or maybe browser DOM) doesn't support adding children to <script>, so we can't use a nice ComplexPanel for this. The hackery necessary to do it is wrapped up in ServerFBMLPanel.render. Phew.