Instead of having the layout perform "business" logic, have the games page request the facebook friends bar using Frame.openBottomFrame (Frame should really be called Shell or Site or MainFrame or something, but I'll save that for another day). The FrameEntryPoint then handles this by telling the layout to set its bottom content. This way, the layout only needs to do the layout and not handle a bunch of messy logic to do with pages and tokens etc.
Also got rid of isTitlePage. This doesn't make sense because whether a page sets the title or not depends on whether it is in the "main" frame or not rather than exactly what page it is. This is accomplished by checking the frameId parameter when Frame.Call.SET_TITLE is requested. Also also added final constants for the main frame and bottom frames.

Also also also added some additional logic to frameCall behavior based on the frame id, though only the main frame's page will be calling this in production.