Checkpoint on the "relisting bug". It's more complicated than I thought because the CatalogRecord is presently deleted. The problem is that if you list item, it gets delisted (either by you, an admin, or automatically (for derived items)), when you relist it a new CatalogRecord is created with a new master item copy.
This is fucked because it breaks the chain of updating clones, it breaks
favorites, it breaks all kinds of things. If you want something to
be a different master item, you upload a new master item, you don't
use an undocumented side-effect that is the opposite of what we
usually want.

But, the fix isn't in yet. A lot of things check catalogId directly
instead of using the methods on Item or ItemRecord, I'll come back
to this after I finish up a mochi thing.