Indiana Greenwell and the Hare-Raising Adventure! In this episode, he falls into a pit filled with... rabbit turds! The dreaded rabbit hole! Because see, when we're in a hurry, it's ok to hack things, right? So then we waste a fucking pile of time later. Stick that carrot up your ass and start smoking.
Oh boy, I'm going to go home and tell the girlfriend that I made
a text input widget get disabled today. SHE WILL BE SO PROUD.

- Cleaned up where the chat input color got set, because I also needed
to disable it if you're in a "read-only" channel.
- Then, after slamming my dick in a door for quite a while, I discovered
that the first place that gets set up sets up a ChatTab connected to
a MsoyChatChannel with the RoomName. But after that, the fucking
text on the tab is changed, with no change to the channel or
RoomName. Because when chat is input on this channel, other fucking
code says "oh, it's PLACE_CHAT, just shoot that out on the current
room object" and so the fact that the RoomName has a stale name
and a stale sceneId don't matter. Except to Mr. Maintainer Coder,
who thinks that code works the way it looks like it works. Fuck.