Checkpoint. Well, unbelievable. It seems that unloadAndStop() can be circumvented. I've found some people talking about this issue, but no official bug on adobe's site, that I've found. In any case, my test swf can do it, and I'm not even sure how it is.
I guess... until we force everyone to use flash 10, we have this problem
anyway because unloadAndStop() doesn't exist in 9.

So... cack. I was hoping to be able to unload this thing completely
before putting our client on the stage, but if that can't be done
then we need to load this in a security boundary. I was really hoping
to avoid having to create some sort of fucking API so that the thing
could find out how big the stage was. I'm sick of knitting mittens
and forcing flash developers to wear them.

This also aids my suspended investingations of stage-accessible games.
If we allow anything access to the stage, then I think we have to kill
that flash application afterward, otherwise it can do nasty things.
For instance, my test code pushes all the buttons and changes the
text on anything it finds whenever you type a "q" anywhere in the client.
So it sniffs your chat, and fucks everything up royally.