Bugfix: fix launchers so that they actually know they are for AVRGs. You never saw that you say? Because when we have GWT we drop that info anyway, route the gameId through the URL and back, and end up correcting the mistake. (Then we revert the URL to point to the scene again. It's so funny it hurts, only without the funny part.)
I was going to just make the fucking Launcher editor just fill
this info in, but it doesn't know it. Nor does the page that
loads the Launcher editor. 

Let us refer to the ancient texts:

   "Thou shalt put all ye annoyinge info into anchor tokens,
    so that thy camel's balls can be bookmarked, every haire
    individulee, but the devil's search engineses shall be thwarted."
   Thus spake the Bayne.
                 
It was too annoying to add another freaking parameter onto all
these pages, so instead I just jam the info in when the item is
sent to the server to be saved.

TO-PROBABLY-DO:
- fix existing launchers
- update the launchers for a game when the avrg-ness of the game changes.
It's unlikely, but on the other hand, anyone editing the game like
that is on whirled.com and wouldn't see the problem, and so wouldn't
know to recreate their launchers.

TODO:
- stop fucking working different ways, and simplify, so that can test
our own code without having to try a matrix of different ways of 
doing it.
- just make the URL www.whirled.com everywhere, EVERYWHERE.
And give people a "share" or "link here" button that gives them
a /go/ url.